refactor: restructure proposal components for better organization
- Moved ProposalCard and related components into `proposal/card/`
- Moved comment-related components into `proposal/comments/`
- Updated imports in affected files (e.g., `ProposalDetails`, batch-vote cards)
- Created new `proposal/card/index.ts` for cleaner exports
