+++ lib/components/proposal/card/ViewMoreButton.tsx
@@ -1,7 +1,7 @@
 import { Link, router } from "expo-router";
-import { IconViewMore } from "../icons/shared";
-import ThemedButton from "../ThemedButton";
-import { ThemedText } from "../ThemedText";
+import { IconViewMore } from "../../icons/shared";
+import ThemedButton from "../../ThemedButton";
+import { ThemedText } from "../../ThemedText";
 import { StyleSheet } from "react-native";
 import { Post } from "@/lib/types";
 import { useProposalStore } from "@/lib/store/proposalStore";
