+++ lib/components/proposal/card/actions/index.ts
