+++ lib/components/proposal/card/body/ProposalBody.tsx
