+++ lib/components/proposal/card/body/TimeDisplay.tsx
