+++ lib/components/proposal/card/body/index.ts
