+++ lib/components/proposal/card/header/index.ts
