+++ lib/components/proposal/card/index.ts
@@ -1,0 +1,3 @@
+import { ProposalCard, ProposalCardSkeleton } from "./ProposalCard";
+
+export { ProposalCard, ProposalCardSkeleton }
