+++ lib/components/proposal/comments/CommentCard.tsx
@@ -2,12 +2,12 @@ import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { StyleSheet, View } from "react-native";
 import debounce from "lodash/debounce";
 import { ICommentResponse, EReaction } from "@/lib/types";
-import { UserAvatar } from "../shared";
-import ThemedButton from "../ThemedButton";
-import { IconLike, IconDislike, IconComment } from "../icons/shared";
-import { ThemedText } from "../ThemedText";
-import VerticalSeprator from "../shared/VerticalSeprator";
-import StackedAvatars from "./StackedAvatars";
+import { UserAvatar } from "../../shared";
+import ThemedButton from "../../ThemedButton";
+import { IconLike, IconDislike, IconComment } from "../../icons/shared";
+import { ThemedText } from "../../ThemedText";
+import VerticalSeprator from "../../shared/VerticalSeprator";
+import StackedAvatars from "../../feed/StackedAvatars";
 import { extractUniqueChildrenAvatars } from "@/lib/util/commentUtil";
 import CommentBox from "./CommentBox";
 import useAddCommentReaction from "@/lib/net/queries/actions/useAddCommentReaction";
