+++ lib/components/proposal/comments/index.ts
@@ -2,4 +2,4 @@ import CommentBox from "./CommentBox";
 import CommentCard from "./CommentCard";
 import CommentList from "./CommentList";
 
-export { CommentBox, CommentCard, CommentList } 
+export { CommentBox, CommentCard, CommentList }
