+++ lib/components/proposal/index.ts
@@ -1,5 +1,6 @@
-import { ProposalCard, ProposalCardSkeleton } from "./ProposalCard";
+
 import ProposalContentSummary from "./ProposalSummaryCard";
 import ProposalDetails from "./ProposalDetails";
+import { ProposalCard, ProposalCardSkeleton } from "./card";
 
 export { ProposalCard, ProposalCardSkeleton, ProposalContentSummary, ProposalDetails };
